🔧 fix(coinbase): resolve review comment, moving constants to single file
